test(cli): split the db sign ref-advancement tests into their own file
db-sign.test.ts had grown to 610 lines. The ref advancement group and the
contract-ref test move to db-sign.ref-advancement.test.ts; the harness, fake
client and fixture helpers both files share move to db-sign-fixtures.ts.
